addNotice - fix

+++ src/middlewares/validatinFileType.js
@@ -3,16 +3,17 @@ const fs = require("fs");
 
 const validatinFileType = async (req, res, next) => {
   try {
-    const { path: temporaryName } = req.file;
+    if (req.file) {
+      const { path: temporaryName } = req.file;
+      const [, extension] = temporaryName?.split(".");
 
-    const [, extension] = temporaryName?.split(".");
-
-    if (
-      extension.toLowerCase() !== "jpg" &&
-      extension.toLowerCase() !== "png"
-    ) {
-      await fs.unlinkSync(temporaryName);
-      next(new ValidationError("file must be '.jpg' or '.png'"));
+      if (
+        extension.toLowerCase() !== "jpg" &&
+        extension.toLowerCase() !== "png"
+      ) {
+        await fs.unlinkSync(temporaryName);
+        next(new ValidationError("file must be '.jpg' or '.png'"));
+      }
     }
 
     next();
